consolidated interfaces further
