--- sceval_frontend/src/components/entities/Home.ts
@@ -1,6 +1,0 @@
-export default interface Home {
-    creationTime: string;
-    id: number;
-    name: string;
-    projectId: number;
-  }
