+++ sceval_frontend/src/components/entities/SensorModule.ts
@@ -29,14 +29,3 @@ export interface SensorDataBundle {
     type: string;
     TSDBData: TimeSeriesData;
 }
-export interface Event {
-  eventType: string;
-  eventData: any;
-  timestamp: string;
-  homeId: number;
-  originDeviceId: number;
-  originDeviceClass: string;
-  originDeviceIp: string;
-  originProjectKeyId: string;
-  originProjectKeyName: string;
-}
